backend/foodgram_project_react/requirements.txt
asgiref==3.2.10
Django==2.2.16
django-filter==21.1
djangorestframework==3.12.4
djoser==2.1.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
python-dotenv==0.19.2
pytz==2022.1
sqlparse==0.3.1
Pillow==9.0.1